tweaked channel functionality

+++ src/components/channel/channelComponent.js
@@ -11,6 +11,9 @@ function channelCtrl( $scope, $state, messageService, recorderService, socketFac
         this.isUserMemberInvite = false;
         this.isUserAdminInvite = false;
         this.isUserAnon = true;
+        this.glued = true;
+        this.playing = false;
+        this.hasSong = false;
 
         for ( let i = 0; i < this.channel.members.length; i++ ) {
             if ( this.channel.members[ i ]._id === this.user._id ) {
@@ -70,14 +73,14 @@ function channelCtrl( $scope, $state, messageService, recorderService, socketFac
       }
   };
 
-  this.togglePlay = () => {
-    if( !this.playing ){
-      this.playing = true;
-    }
-    else {
-      this.playing = false;
-    }
-  }
+  // this.togglePlay = () => {
+  //   if( !this.playing ){
+  //     this.playing = true;
+  //   }
+  //   else {
+  //     this.playing = false;
+  //   }
+  // }
 
   this.enterChannel = () => {
       if ( this.channel ) {
@@ -188,8 +191,21 @@ function channelCtrl( $scope, $state, messageService, recorderService, socketFac
 	$scope.playlist = playList.src;
 
 // Record / Play functions
+  this.playSong = () => {
+    this.wavesurfer.play();
+    this.playing = true;
+  }
+
+  this.pauseSong= () => {
+    this.playing = false;
+    this.wavesurfer.pause();
+  }
 
   this.startRecording = () => {
+    if ( this.wavesurfer ){
+      this.wavesurfer.destroy();
+    }
+
     this.wavesurfer = WaveSurfer.create( {
       container: "#waveform"
       , waveColor: "#F46036"
@@ -211,7 +227,7 @@ function channelCtrl( $scope, $state, messageService, recorderService, socketFac
     this.wavesurfer.destroy();
   };
 
-  this.playSong = ( song ) => {
+  this.sendSong = ( song ) => {
     if ( this.wavesurfer ){
       this.wavesurfer.destroy();
     }
@@ -227,7 +243,10 @@ function channelCtrl( $scope, $state, messageService, recorderService, socketFac
     this.wavesurfer.load( song );
   }
 
-  this.deleteRecording = ( recording, channelId ) => {
+  this.deleteRecording = ( recording, channelId, userId ) => {
+    if ( recording.createdBy._id !== userId ){
+        return alert("You can only delete your own recordings.")
+    }
     recordingService.deleteRecording( recording, channelId );
   }
 
@@ -243,13 +262,15 @@ function channelCtrl( $scope, $state, messageService, recorderService, socketFac
 
     this.uploadAndSaveRecording = () => {
       recorderService.uploadRecordingToS3( this.recordingData, this.user._id, this.channel._id );
+      this.wavesurfer.destroy();
+      console.log("Recording Data" + this.recordingData)
     };
 	} );
 
-    this.deletePreview = () => {
-      // this.wavesurfer.destroy();
-      $state.reload();
-    }
+    // this.deletePreview = () => {
+    //   // this.wavesurfer.destroy();
+    //   $state.reload();
+    // }
 
   socketFactory.on( "get S3 data", data => {
     console.log(data);
